+++ README.md
@@ -1,136 +1,114 @@
-# Sweetbreath
+# Sweet Breath 甜之呼吸 - 網路甜點工作室
 
-[Sweetbreath 甜の呼吸 網路甜點工作室](http://sweetbreath.tw/)
+- [網站 DEMO](http://sweetbreath.tw/)
 
 ![](./img/localhost-index02-20210106.png)
 
 ## 專案介紹
 
-本專案為 Sweet Breath 甜の呼吸甜點購物車前端與後端原始碼，主要開發技術為 React Hooks、
-styled-components、Express、Sequelize、MySQL。
+本專案為 Sweet Breath 甜之呼吸 - 網路甜點工作室的前端原始碼，採用前端技術為 React Hooks，搭配後端技術
+Express 和 Sequelize 進行開發。
 
-使用者相關功能
+## 專案展示
 
-`商品模組`
+### 使用者相關功能
 
-- 商品分類檢視：商品分類以外，可建立特殊商品分類，例如人氣商品。
-- 瀏覽所有商品：依商品分類呈現。
-- 瀏覽單一商品：瀏覽商品名稱、商品介紹、商品規格、商品庫存、選擇商品數量加入購物車。
-
-`購物模組`
-
-- 購物車：選擇商品數量加入購物車，顯示訂單摘要。
-- 結帳：導向登入頁面，登入後輸入配送資料並檢視輸入格式，確認訂單總額後提交訂單。
-
-`會員模組`
-
-- 會員註冊：必須輸入全名、帳號、電子郵件、密碼，檢視符合輸入格式即可註冊成功。
-- 會員登入：必須輸入帳號、密碼，即可登入。登入後依會員權限顯示一般權限選單／管理權限選單。
-- 會員專區：會員登入後可瀏覽個人資料、消費紀錄。個人資料可修改全名、電子郵件、生日、地址，為統一修改模式，全名、電子郵件必填，其他未填欄位於送出後清除原本資料。消費紀錄可瀏覽全部訂單，並顯示訂單狀態、訂單明細，點擊商品圖片、商品名稱可瀏覽商品介紹。
-
-`基本模組`
-
-- 首頁：主打活動輪播區、人氣商品。
-- 關於我們：瀏覽網站架設理念。
-- 最新消息：瀏覽最新消息列表。
-- 聯絡我們：瀏覽店家聯絡資料，瀏覽 Google Maps 實體地址、聯絡表單。
-
-管理者相關功能
-
-`商品管理模組`
-
-- 依商品分類瀏覽所有商品。
-- 編輯商品：可編輯商品資訊，包括商品名稱、圖片網址、狀態、分類、商品介紹。
-- 刪除商品：後端處理為變更商品狀態，仍保留商品資訊。
-- 新增商品：可新增商品資訊，包括商品資訊、商品規格，商品規格包括規格名稱、原價、特價、庫存。
-
-`分類管理模組`
-
-- 瀏覽所有商品分類名稱，顯示各類商品數量。
-- 編輯／刪除／新增分類。
-- 刪除分類：刪除商品：後端處理為變更分類狀態，仍保留分類名稱。
-
-`權限管理模組`
-
-- 瀏覽所有帳號，可依使用者 id 或名稱搜尋特定帳號。
-- 分為一般權限、管理權限、停權。
-- 停權：後端處理為變更帳號狀態，仍保留帳號資訊。
-
-`訂單管理模組`
-
-- 瀏覽所有訂單明細，依訂單狀態顯示訂單。
-- 管理訂單狀態，依付款進度、出貨進度調整訂單狀態，或取消處理中訂單。
-- 取消訂單：後端處理為變更訂單狀態，仍保留訂單資訊。
+#### `頁面架構`
 
 ![](./img/userflow-client-v3-20210104.png)
 
+#### `產品模組`
+
+- 瀏覽所有產品：可依照分類檢視產品。
+- 瀏覽單一產品：可瀏覽產品詳細資訊、產品規格與庫存，並可選擇產品數量加入購物車。
+
+![](./img/)
+
+#### `購物模組`
+
+- 購物車系統：顯示加入購物車的商品，並顯示訂單摘要。
+- 結帳功能：導向登入頁面，登入後輸入配送資料，確認訂單資訊和總額後，即可建立訂單。
+
+![](./img/)
+
+#### `會員模組`
+
+- 會員身分驗證功能：登入後可查看或編輯個人相關資訊。
+- 查看訂單：可在消費記錄查看已建立的訂單資訊。
+
+![](./img/)
+
+#### `基本模組`
+
+- 首頁：主打活動輪播區，以及顯示人氣產品。
+- 關於我們：瀏覽網站架設理念。
+- 最新消息：瀏覽最新消息列表。
+- 聯絡我們：瀏覽店家相關聯絡資訊、Google Maps 實體地址。
+
+#### 管理者相關功能
+
+#### `頁面架構`
+
 ![](./img/userflow-admin-v3-20210104.png)
 
-## 專案後端
+#### `產品管理模組`
 
-[sweetbreath-backend](https://github.com/ivymuchacha/sweetbreath-backend)：主要開發技術為 Express、Sequelize、MySQL。
+- 依分類管理：可新增或編輯產品分類。
+- 依規格管理：可新增或編輯產品資訊、產品規格。
+
+![](./img/)
+
+#### `權限管理模組`
+
+- 會員搜尋功能：可瀏覽所有帳號，依照使用者 id 或名稱搜尋特定帳號。
+- 會員管理功能：可變更會員權限狀態，分為一般權限、管理權限、停權。
+
+![](./img/)
+
+#### `訂單管理模組`
+
+- 瀏覽所有訂單：可查看所有訂單明細，並依照訂單狀態顯示訂單進度。
+- 訂單管理功能：可依照付款進度、出貨進度調整訂單狀態，或取消處理中訂單。
+
+![](./img/)
 
 ## 專案執行
 
-`npm install`
+1. 執行 `npm install` 安裝專案所需套件。
 
-安裝專案所需套件。
+2. 執行 `npm run start`，在 http://localhost:3000 啟動專案。
 
-`npm run dev`
+3. 執行 `npm run dev`，在 http://localhost:5000 啟動專案後端。
 
-http://localhost:3000 啟動專案。  
-http://localhost:5000 啟動專案後端。
+4. 執行 `npm build`，在 build 資料夾建立專案 production 版本。
 
-`npm build`
+5. 執行 `npm deploy`，在 GitHub Pages 部屬專案網站。
 
-資料夾 build 建立專案 production 版本。
+## 專案前端技術
 
-`npm deploy`
-
-GitHub Pages 部屬專案網站。
-
-## 專案技術
-
-`前端框架與套件`
+#### `框架`
 
 - [React Hooks](https://reactjs.org/)
-- [Create React App](https://create-react-app.dev/)
 - [React DOM](https://www.npmjs.com/package/react-dom)
+
+#### `第三方套件`
+
 - [React Router DOM](https://reactrouter.com/)
-- [PropTypes](https://www.npmjs.com/package/prop-types)
 - [styled-components](https://styled-components.com/)
+- [PropTypes](https://www.npmjs.com/package/prop-types)
 - [ESLint](https://eslint.org/)
 - [Prettier](https://prettier.io/)
 - [Google Map React](https://github.com/google-map-react/google-map-react)
 
-`後端框架與套件`
-
-- [Express](https://expressjs.com/)
-- [Sequelize](https://sequelize.org/)
-- [MySQL](https://www.mysql.com/)
-
-`第三方 API`
+#### `第三方 API`
 
 - [Google Cloud Platform](https://cloud.google.com/)
-- [Imgur](https://imgur.com/)
 
-## 專案展示
+## 專案後端技術
 
-### 商品模組
+Sweet Breath 甜之呼吸的後端原始碼，採用 Express 和 Sequelize 開發，生成 API 與前端串連。
 
-![](./img/)
-
-### 購物模組
-
-![](./img/)
-
-### 會員模組
-
-![](./img/)
-
-### 後臺管理模組
-
-![](./img/)
+- 專案連結：[sweetbreath-backend](https://github.com/ivymuchacha/sweetbreath-backend)
 
 ## 專案授權
