--- src/pages/admin/AdminNewsPage/AdminNewsPage.js
@@ -1,3 +1,0 @@
-export default function AdminNewsPage() {
-  return <div>ADMIN_NEWS</div>;
-}
