--- src/pages/admin/AdminNewsPage/index.js
@@ -1,1 +1,0 @@
-export { default } from "./AdminNewsPage";
