+++ src/pages/index.js
@@ -1,7 +1,6 @@
 import HomePage from "./user/HomePage";
 import AboutPage from "./user/AboutPage";
 import LoginPage from "./user/LoginPage";
-import LogoutPage from "./user/LogoutPage";
 import RegisterPage from "./user/RegisterPage";
 import ProductListPage from "./user/ProductListPage";
 import ProductPage from "./user/ProductPage";
@@ -16,7 +15,6 @@ import AdminProductPage from "./admin/AdminProductPage";
 import AdminEditProductPage from "./admin/AdminEditProductPage";
 import AdminMemberPage from "./admin/AdminMemberPage";
 import AdminOrderListPage from "./admin/AdminOrderListPage";
-import AdminNewsPage from "./admin/AdminNewsPage";
 import AdminCategoryPage from "./admin/AdminCategory";
 import OrderListPage from "./user/OrderListPage";
 
@@ -24,7 +22,6 @@ export {
   HomePage,
   AboutPage,
   LoginPage,
-  LogoutPage,
   RegisterPage,
   ProductListPage,
   ProductPage,
@@ -39,7 +36,6 @@ export {
   AdminEditProductPage,
   AdminMemberPage,
   AdminOrderListPage,
-  AdminNewsPage,
   AdminCategoryPage,
   OrderListPage,
 };
