--- src/pages/user/LogoutPage/LogoutPage.js
@@ -1,3 +1,0 @@
-export default function LogoutPage() {
-  return <div>LOGOUT</div>;
-}
