--- src/pages/user/LogoutPage/index.js
@@ -1,1 +1,0 @@
-export { default } from "./LogoutPage";
